- Rearranged endpoint file structure. - Added subprocess calling helper to util. - Improved logging.
