+++ btrfs_backup/endpoint/__init__.py
@@ -1,0 +1,3 @@
+from .local import LocalEndpoint
+from .ssh import SSHEndpoint
+from .shell import ShellEndpoint
